feat: search/sort/edit/late/supplies improvements
- Add ConflictError type returning conflicting apt on 409
- Add PATCH /appointments/:id/late endpoint (SetLate service method)
- Supply model: add unit, quantity_grams, total_cost, cost_per_unit (computed)
- ServiceSupply.Quantity: int → float64 for gram precision
- Supply.Quantity: int → float64
- DeductQuantity: float64
- supply service: compute cost_per_unit in GetAll/GetByType/Create

+++ internal/repository/service_supply.go
@@ -42,6 +42,6 @@ func (r *ServiceSupplyRepository) DeleteByServiceID(serviceID uint) error {
 	return r.db.Where("service_id = ?", serviceID).Delete(&model.ServiceSupply{}).Error
 }
 
-func (r *ServiceSupplyRepository) UpdateQuantity(id uint, quantity int) error {
+func (r *ServiceSupplyRepository) UpdateQuantity(id uint, quantity float64) error {
 	return r.db.Model(&model.ServiceSupply{}).Where("id = ?", id).Update("quantity", quantity).Error
 }
